ANDROID: KVM: arm64: Do not use kvm.created_vcpus to track the number of pinned vcpus
Do not use variable to reflect something it wasn't intended to reflect, i.e.,
number of created vcpus vs number of vcpus pinned so far.

Consolidate pinning and error handling to the same level to make
code more readable.

Ensure that the donated pgd is big enough for all vcpus.

Bug: 220830416
Bug: 216808671
Signed-off-by: Fuad Tabba <tabba@google.com>
Change-Id: Ibf41a93bb1175e59b3ab82d2f735f25505d2892a